[rename] 공통 response 폴더 생성 및 파일 이동, ResponseProperties 파일 생성
- 공통적으로 사용하는 response 담을 폴더 생성 및 파일 이동
