+++ myhands/src/main/java/tabom/myhands/common/properties/ResponseProperties.java
@@ -1,0 +1,13 @@
+package tabom.myhands.common.properties;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Getter
+@RequiredArgsConstructor
+@ConfigurationProperties(prefix = "response")
+public class ResponseProperties {
+    private final String success;
+    private final String fail;
+}
